More menu stuff, plus switch to object on creation

+++ src/main.cpp
@@ -279,15 +279,15 @@ void toggle_cameras (int val){
 }
 
 void handle_objects (int val){
-
+    int index = -1;
     switch ((Object)val){
         case Object::Box : {
             
             // Create dynamic box object
             boxhdl *box = new boxhdl(1.0, 1.0, 1.0);
             scene.objects.push_back(box);
-            
-            current_objects->add_item((int)scene.objects.size()-1, "Box");
+            index = (int)scene.objects.size()-1;
+            current_objects->add_item(index, "Box");
             break;
         }
         case Object::Cylinder : {
@@ -295,8 +295,8 @@ void handle_objects (int val){
             // Create dynamic sphere object
             cylinderhdl *cylinder = new cylinderhdl(1.0, 4.0, 10.0);
             scene.objects.push_back(cylinder);
-            
-            current_objects->add_item((int)scene.objects.size()-1, "Cylinder");
+            index = (int)scene.objects.size()-1;
+            current_objects->add_item(index, "Cylinder");
             break;
         }
         case Object::Sphere : {
@@ -304,8 +304,8 @@ void handle_objects (int val){
             // Create dynamic sphere object
             spherehdl *sphere = new spherehdl(1.0, 8.0, 16.0);
             scene.objects.push_back(sphere);
-            
-            current_objects->add_item((int)scene.objects.size()-1, "Sphere");
+            index = (int)scene.objects.size()-1;
+            current_objects->add_item(index, "Sphere");
             break;
         }
         case Object::Pyramid : {
@@ -313,8 +313,8 @@ void handle_objects (int val){
             // Create dynamic sphere object
             pyramidhdl *pyramid = new pyramidhdl(1.0, 4.0, 12.0);
             scene.objects.push_back(pyramid);
-            
-            current_objects->add_item((int)scene.objects.size()-1, "Pyramid");
+            index = (int)scene.objects.size()-1;
+            current_objects->add_item(index, "Pyramid");
             break;
         }
         case Object::Model :
@@ -324,11 +324,13 @@ void handle_objects (int val){
             
             break;
     }
+    current_objects->set_int_val(index);
     glutPostRedisplay();
 }
 
 void handle_cameras (int val){
-    
+    int index = -1;
+
     switch ((Camera)val){
         case Camera::Fovy :
             
@@ -357,8 +359,9 @@ void handle_cameras (int val){
         case Camera::Ortho : {
             camerahdl *camera = new orthohdl;
             scene.cameras.push_back(camera);
-            scene.active_camera = (int)scene.cameras.size() - 1;
-            current_cameras->add_item((int)scene.cameras.size()-1, "Ortho");
+            index = (int)scene.cameras.size() - 1;
+            scene.active_camera = index;
+            current_cameras->add_item(index, "Ortho");
 
             camera->view(scene.canvas);
             break;
@@ -366,8 +369,9 @@ void handle_cameras (int val){
         case Camera::Frustum : {
             camerahdl *camera = new frustumhdl;
             scene.cameras.push_back(camera);
-            scene.active_camera = (int)scene.cameras.size() - 1;
-            current_cameras->add_item((int)scene.cameras.size()-1, "Frustum");
+            index = (int)scene.cameras.size() - 1;
+            scene.active_camera = index;
+            current_cameras->add_item(index, "Frustum");
 
             camera->view(scene.canvas);
             break;
@@ -375,8 +379,9 @@ void handle_cameras (int val){
         case Camera::Perspective : {
             camerahdl *camera = new perspectivehdl;
             scene.cameras.push_back(camera);
-            scene.active_camera = (int)scene.cameras.size() - 1;
-            current_cameras->add_item((int)scene.cameras.size()-1, "Perspective");
+            index = (int)scene.cameras.size() - 1;
+            scene.active_camera = index;
+            current_cameras->add_item(index, "Perspective");
 
             camera->view(scene.canvas);
             break;
@@ -385,6 +390,7 @@ void handle_cameras (int val){
             
             break;
     }
+    current_cameras->set_int_val(index);
     glutPostRedisplay();
 }
 
@@ -454,15 +460,18 @@ void handle_delete(int val)
 {
     if (val == 0) {
         int obj_ind = current_objects->get_int_val();
-        scene.objects[obj_ind] = NULL; // TODO: list will keep getting bigger if you only set pointer to null
-        current_objects->delete_item(obj_ind);
-        
-        current_objects->set_int_val(0);
-        
+        if (obj_ind < scene.objects.size() && obj_ind >= 0) {
+            scene.objects[obj_ind] = NULL; // TODO: list will keep getting bigger if you only set pointer to null
+            current_objects->delete_item(obj_ind);
+            current_objects->set_int_val(-1);
+        }
     } else if (val == 1) {
         int obj_ind = current_cameras->get_int_val();
-        scene.cameras[obj_ind] = NULL; // TODO: list will keep getting bigger if you only set pointer to null
-        current_cameras->delete_item(obj_ind);
+        if (obj_ind < scene.cameras.size() && obj_ind >= 0) {
+            scene.cameras[obj_ind] = NULL; // TODO: list will keep getting bigger if you only set pointer to null
+            current_cameras->delete_item(obj_ind);
+            current_cameras->set_int_val(-1);
+        }
     }
     glutPostRedisplay();
 }
@@ -562,40 +571,41 @@ void create_menu()
  */
 void setup_glui() {
     glui = GLUI_Master.create_glui("Controls",0,800,0);
+    GLUI_Panel *scene_panel = glui->add_panel("Current Scene");
     
-    glui->add_statictext("Current Scene");
-    current_objects = glui->add_listbox("Objects");
-    glui->add_button("Delete", 0, handle_delete);
+    current_objects = glui->add_listbox_to_panel(scene_panel, "Objects");
+    current_objects->add_item(-1, "");
+    glui->add_button_to_panel(scene_panel, "Delete", 0, handle_delete);
 
-    glui->add_separator();
-    current_cameras = glui->add_listbox("Cameras");
-    glui->add_button("Delete", 1, handle_delete);
+    glui->add_separator_to_panel(scene_panel);
+    current_cameras = glui->add_listbox_to_panel(scene_panel, "Cameras");
+    current_cameras->add_item(-1, "");
+    glui->add_button_to_panel(scene_panel, "Delete", 1, handle_delete);
     
-    glui->add_separator();
+    glui->add_separator_to_panel(scene_panel);
 
-    GLUI_Listbox *list_polygon = glui->add_listbox("Polygon");
+    GLUI_Listbox *list_polygon = glui->add_listbox_to_panel(scene_panel, "Polygon");
     list_polygon->add_item((int)Polygon::Line,"Line");
     list_polygon->add_item((int)Polygon::Point,"Point");
-    GLUI_Listbox *list_manipulate = glui->add_listbox("Manipulation", &current_manipulation);
+    GLUI_Listbox *list_manipulate = glui->add_listbox_to_panel(scene_panel, "Manipulation", &current_manipulation);
     list_manipulate->add_item(manipulate::translate,"Translate");
     list_manipulate->add_item(manipulate::rotate,   "Rotate");
     list_manipulate->add_item(manipulate::scale,    "Scale");
-    glui->add_checkbox("Draw Cameras", NULL, 1, toggle_cameras);
-    glui->add_separator();
-
+    glui->add_separator_to_panel(scene_panel);
+    glui->add_checkbox_to_panel(scene_panel, "Draw Cameras", NULL, 1, toggle_cameras);
 
     glui->add_column(true);
-    glui->add_statictext("Create Object");
-    glui->add_button("Box",         (int)Object::Box,       handle_objects);
-    glui->add_button("Cylinder",    (int)Object::Cylinder,  handle_objects);
-    glui->add_button("Sphere",      (int)Object::Sphere,    handle_objects);
-    glui->add_button("Pyramid",     (int)Object::Pyramid,   handle_objects);
-    glui->add_button("Model",       (int)Object::Model,     handle_objects);
+    GLUI_Panel *obj_panel = glui->add_panel("Create Object");
+    glui->add_button_to_panel(obj_panel,    "Box",         (int)Object::Box,       handle_objects);
+    glui->add_button_to_panel(obj_panel,    "Cylinder",    (int)Object::Cylinder,  handle_objects);
+    glui->add_button_to_panel(obj_panel,    "Sphere",      (int)Object::Sphere,    handle_objects);
+    glui->add_button_to_panel(obj_panel,    "Pyramid",     (int)Object::Pyramid,   handle_objects);
+    glui->add_button_to_panel(obj_panel,    "Model",       (int)Object::Model,     handle_objects);
     
-    glui->add_statictext("Create Camera");
-    glui->add_button("Ortho",       (int)Camera::Ortho,         handle_cameras);
-    glui->add_button("Frustum",     (int)Camera::Frustum,       handle_cameras);
-    glui->add_button("Perspective", (int)Camera::Perspective,   handle_cameras);
+    GLUI_Panel *cam_panel = glui->add_panel("Create Camera");
+    glui->add_button_to_panel(cam_panel,    "Ortho",       (int)Camera::Ortho,         handle_cameras);
+    glui->add_button_to_panel(cam_panel,    "Frustum",     (int)Camera::Frustum,       handle_cameras);
+    glui->add_button_to_panel(cam_panel,    "Perspective", (int)Camera::Perspective,   handle_cameras);
     
     glui->add_separator();
     glui->add_button("Quit", 0, exit);
@@ -650,12 +660,7 @@ int main(int argc, char **argv)
     setup_glui();
     
     // Create camera
-    camerahdl *camera = new orthohdl;
-    scene.cameras.push_back(camera);
-    scene.active_camera = (int)scene.cameras.size() - 1;
-    current_cameras->add_item((int)scene.cameras.size()-1, "Ortho");
-    
-    camera->view(scene.canvas);
+    handle_cameras((int)Camera::Ortho);
     
     // Start GLUT loop
 	glutMainLoop();
